refactor for downstream
